move InteractiveUtils to stdlib
