+++ base/exports.jl
@@ -761,7 +761,6 @@ export
     promote,
     promote_rule,
     promote_type,
-    subtypes,
     instances,
     supertype,
     typeintersect,
@@ -777,26 +776,17 @@ export
     parse,
 
 # help and reflection
-    apropos,
-    edit,
     code_typed,
-    code_warntype,
     code_lowered,
-    code_llvm,
-    code_native,
     fullname,
     functionloc,
     isconst,
     isinteractive,
-    less,
     hasmethod,
     methods,
-    methodswith,
     nameof,
     parentmodule,
     names,
-    varinfo,
-    versioninfo,
     which,
     @isdefined,
 
@@ -1032,17 +1022,6 @@ export
     @timev,
     @elapsed,
     @allocated,
-
-    # reflection
-    @which,
-    @edit,
-    @functionloc,
-    @less,
-    @code_typed,
-    @code_warntype,
-    @code_lowered,
-    @code_llvm,
-    @code_native,
 
     # tasks
     @schedule,
