+++ base/precompile.jl
@@ -608,8 +608,6 @@ precompile(Tuple{typeof(Base._setindex!), Base.Dict{Union{DataType, typeof(Type)
 precompile(Tuple{typeof(Base.setindex!), Base.Dict{Union{DataType, typeof(Type)}, Nothing}, Nothing, DataType})
 precompile(Tuple{typeof(Base.setindex!), Base.Dict{Union{DataType, typeof(Type)}, Nothing}, Nothing, typeof(Type)})
 precompile(Tuple{typeof(Base.rehash!), Base.Dict{Union{DataType, typeof(Type)}, Nothing}, Int64})
-precompile(Tuple{typeof(Base._subtypes), Module, DataType, Base.Set{Union{DataType, typeof(Type)}}, Base.Set{Module}})
-precompile(Tuple{typeof(Base._subtypes), Module, DataType})
 precompile(Tuple{typeof(Base.isempty), Array{Int64, 1}})
 precompile(Tuple{typeof(Base.Iterators.zip), Array{Int64, 1}, SimpleVector})
 precompile(Tuple{typeof(Base.Iterators.zip), Array{Symbol, 1}, SimpleVector})
