+++ base/reflection.jl
@@ -584,74 +584,6 @@ julia> instances(Color)
 """
 function instances end
 
-# subtypes
-function _subtypes(m::Module, x::Union{DataType,UnionAll},
-                   sts=Set{Union{DataType,UnionAll}}(), visited=Set{Module}())
-    push!(visited, m)
-    xt = unwrap_unionall(x)
-    if !isa(xt, DataType)
-        return sts
-    end
-    xt = xt::DataType
-    for s in names(m, all = true)
-        if isdefined(m, s) && !isdeprecated(m, s)
-            t = getfield(m, s)
-            if isa(t, DataType)
-                t = t::DataType
-                if t.name.name === s && supertype(t).name == xt.name
-                    ti = typeintersect(t, x)
-                    ti != Bottom && push!(sts, ti)
-                end
-            elseif isa(t, UnionAll)
-                t = t::UnionAll
-                tt = unwrap_unionall(t)
-                isa(tt, DataType) || continue
-                tt = tt::DataType
-                if tt.name.name === s && supertype(tt).name == xt.name
-                    ti = typeintersect(t, x)
-                    ti != Bottom && push!(sts, ti)
-                end
-            elseif isa(t, Module)
-                t = t::Module
-                in(t, visited) || _subtypes(t, x, sts, visited)
-            end
-        end
-    end
-    return sts
-end
-
-function _subtypes_in(mods::Array, x::Union{DataType,UnionAll})
-    if !isabstracttype(x)
-        # Fast path
-        return Union{DataType,UnionAll}[]
-    end
-    sts = Set{Union{DataType,UnionAll}}()
-    visited = Set{Module}()
-    for m in mods
-        _subtypes(m, x, sts, visited)
-    end
-    return sort!(collect(sts), by=string)
-end
-
-subtypes(m::Module, x::Union{DataType,UnionAll}) = _subtypes_in([m], x)
-
-"""
-    subtypes(T::DataType)
-
-Return a list of immediate subtypes of DataType `T`. Note that all currently loaded subtypes
-are included, including those not visible in the current module.
-
-# Examples
-```jldoctest
-julia> subtypes(Integer)
-3-element Array{Union{DataType, UnionAll},1}:
- Bool
- Signed
- Unsigned
-```
-"""
-subtypes(x::Union{DataType,UnionAll}) = _subtypes_in(loaded_modules_array(), x)
-
 function to_tuple_type(@nospecialize(t))
     @_pure_meta
     if isa(t,Tuple) || isa(t,AbstractArray) || isa(t,SimpleVector)
@@ -858,79 +790,6 @@ struct CodegenParams
             module_setup, module_activation, raise_exception)
 end
 
-# Printing code representations in IR and assembly
-function _dump_function(@nospecialize(f), @nospecialize(t), native::Bool, wrapper::Bool,
-                        strip_ir_metadata::Bool, dump_module::Bool, syntax::Symbol=:att,
-                        optimize::Bool=true, params::CodegenParams=CodegenParams())
-    ccall(:jl_is_in_pure_context, Bool, ()) && error("code reflection cannot be used from generated functions")
-    if isa(f, Core.Builtin)
-        throw(ArgumentError("argument is not a generic function"))
-    end
-    # get the MethodInstance for the method match
-    world = typemax(UInt)
-    meth = which(f, t)
-    t = to_tuple_type(t)
-    tt = signature_type(f, t)
-    (ti, env) = ccall(:jl_type_intersection_with_env, Any, (Any, Any), tt, meth.sig)::SimpleVector
-    meth = func_for_method_checked(meth, ti)
-    linfo = ccall(:jl_specializations_get_linfo, Ref{Core.MethodInstance}, (Any, Any, Any, UInt), meth, ti, env, world)
-    # get the code for it
-    return _dump_function_linfo(linfo, world, native, wrapper, strip_ir_metadata, dump_module, syntax, optimize, params)
-end
-
-function _dump_function_linfo(linfo::Core.MethodInstance, world::UInt, native::Bool, wrapper::Bool,
-                              strip_ir_metadata::Bool, dump_module::Bool, syntax::Symbol=:att,
-                              optimize::Bool=true, params::CodegenParams=CodegenParams())
-    if syntax != :att && syntax != :intel
-        throw(ArgumentError("'syntax' must be either :intel or :att"))
-    end
-    if native
-        llvmf = ccall(:jl_get_llvmf_decl, Ptr{Cvoid}, (Any, UInt, Bool, CodegenParams), linfo, world, wrapper, params)
-    else
-        llvmf = ccall(:jl_get_llvmf_defn, Ptr{Cvoid}, (Any, UInt, Bool, Bool, CodegenParams), linfo, world, wrapper, optimize, params)
-    end
-    if llvmf == C_NULL
-        error("could not compile the specified method")
-    end
-
-    if native
-        str = ccall(:jl_dump_function_asm, Ref{String},
-                    (Ptr{Cvoid}, Cint, Ptr{UInt8}), llvmf, 0, syntax)
-    else
-        str = ccall(:jl_dump_function_ir, Ref{String},
-                    (Ptr{Cvoid}, Bool, Bool), llvmf, strip_ir_metadata, dump_module)
-    end
-
-    # TODO: use jl_is_cacheable_sig instead of isdispatchtuple
-    isdispatchtuple(linfo.specTypes) || (str = "; WARNING: This code may not match what actually runs.\n" * str)
-    return str
-end
-
-"""
-    code_llvm([io=STDOUT,], f, types)
-
-Prints the LLVM bitcodes generated for running the method matching the given generic
-function and type signature to `io`.
-
-All metadata and dbg.* calls are removed from the printed bitcode. Use code_llvm_raw for the full IR.
-"""
-code_llvm(io::IO, @nospecialize(f), @nospecialize(types=Tuple), strip_ir_metadata=true, dump_module=false) =
-    print(io, _dump_function(f, types, false, false, strip_ir_metadata, dump_module))
-code_llvm(@nospecialize(f), @nospecialize(types=Tuple)) = code_llvm(STDOUT, f, types)
-code_llvm_raw(@nospecialize(f), @nospecialize(types=Tuple)) = code_llvm(STDOUT, f, types, false)
-
-"""
-    code_native([io=STDOUT,], f, types; syntax = :att)
-
-Prints the native assembly instructions generated for running the method matching the given
-generic function and type signature to `io`.
-Switch assembly syntax using `syntax` symbol parameter set to `:att` for AT&T syntax or `:intel` for Intel syntax.
-"""
-code_native(io::IO, @nospecialize(f), @nospecialize(types=Tuple); syntax::Symbol = :att) =
-    print(io, _dump_function(f, types, true, false, false, false, syntax))
-code_native(@nospecialize(f), @nospecialize(types=Tuple); syntax::Symbol = :att) = code_native(STDOUT, f, types, syntax = syntax)
-code_native(::IO, ::Any, ::Symbol) = error("illegal code_native call") # resolve ambiguous call
-
 # give a decent error message if we try to instantiate a staged function on non-leaf types
 function func_for_method_checked(m::Method, @nospecialize types)
     if isdefined(m, :generator) && !isdispatchtuple(types)
@@ -1004,13 +863,11 @@ function which(@nospecialize(f), @nospecialize(t))
 end
 
 """
-    which(symbol)
+    which(module, symbol)
 
-Return the module in which the binding for the variable referenced by `symbol` in module `Main` was created.
+Return the module in which the binding for the variable referenced by `symbol` in `module` was created.
 """
-which(s::Symbol) = which_module(Main, s)
-# TODO: making this a method of which() causes a strange error
-function which_module(m::Module, s::Symbol)
+function which(m::Module, s::Symbol)
     if !isdefined(m, s)
         error("\"$s\" is not defined in module $m")
     end
