+++ base/util.jl
@@ -675,18 +675,36 @@ kwdef_val(::Type{T}) where {T<:Integer} = zero(T)
 
 kwdef_val(::Type{T}) where {T} = T()
 
+# testing
 
-function _check_bitarray_consistency(B::BitArray{N}) where N
-    n = length(B)
-    if N ≠ 1
-        all(d ≥ 0 for d in B.dims) || (@warn("Negative d in dims: $(B.dims)"); return false)
-        prod(B.dims) ≠ n && (@warn("Inconsistent dims/len: prod(dims)=$(prod(B.dims)) len=$n"); return false)
+"""
+    Base.runtests(tests=["all"]; ncores=ceil(Int, Sys.CPU_CORES / 2),
+                  exit_on_error=false, [seed])
+
+Run the Julia unit tests listed in `tests`, which can be either a string or an array of
+strings, using `ncores` processors. If `exit_on_error` is `false`, when one test
+fails, all remaining tests in other files will still be run; they are otherwise discarded,
+when `exit_on_error == true`.
+If a seed is provided via the keyword argument, it is used to seed the
+global RNG in the context where the tests are run; otherwise the seed is chosen randomly.
+"""
+function runtests(tests = ["all"]; ncores = ceil(Int, Sys.CPU_CORES / 2),
+                  exit_on_error=false,
+                  seed::Union{BitInteger,Nothing}=nothing)
+    if isa(tests,AbstractString)
+        tests = split(tests)
     end
-    Bc = B.chunks
-    nc = length(Bc)
-    nc == num_bit_chunks(n) || (@warn("Incorrect chunks length for length $n: expected=$(num_bit_chunks(n)) actual=$nc"); return false)
-    n == 0 && return true
-    Bc[end] & _msk_end(n) == Bc[end] || (@warn("Nonzero bits in chunk after `BitArray` end"); return false)
-    return true
+    exit_on_error && push!(tests, "--exit-on-error")
+    seed != nothing && push!(tests, "--seed=0x$(hex(seed % UInt128))") # cast to UInt128 to avoid a minus sign
+    ENV2 = copy(ENV)
+    ENV2["JULIA_CPU_CORES"] = "$ncores"
+    try
+        run(setenv(`$(julia_cmd()) $(joinpath(Sys.BINDIR,
+            Base.DATAROOTDIR, "julia", "test", "runtests.jl")) $tests`, ENV2))
+    catch
+        buf = PipeBuffer()
+        versioninfo(buf)
+        error("A test has failed. Please submit a bug report (https://github.com/JuliaLang/julia/issues)\n" *
+              "including error messages above and the output of versioninfo():\n$(read(buf, String))")
+    end
 end
-
