+++ doc/src/base/base.md
@@ -22,14 +22,7 @@ Base.exit
 Base.quit
 Base.atexit
 Base.isinteractive
-Base.varinfo
 Base.summarysize
-Base.edit(::AbstractString, ::Integer)
-Base.edit(::Any)
-Base.@edit
-Base.less(::AbstractString)
-Base.less(::Any)
-Base.@less
 Base.clipboard(::Any)
 Base.clipboard()
 Base.require
@@ -38,14 +31,9 @@ Base.__precompile__
 Base.include
 Base.include_string
 Base.include_dependency
-Base.Docs.apropos
 Base.which(::Any, ::Any)
-Base.which(::Symbol)
-Base.@which
 Base.methods
-Base.methodswith
 Base.@show
-Base.versioninfo
 ans
 ```
 
@@ -136,7 +124,6 @@ Base.identity
 Base.supertype
 Core.:(<:)
 Base.:(>:)
-Base.subtypes
 Base.typejoin
 Base.typeintersect
 Base.promote_type
@@ -342,7 +329,6 @@ Base.isconst
 Base.nameof(::Function)
 Base.functionloc(::Any, ::Any)
 Base.functionloc(::Method)
-Base.@functionloc
 ```
 
 ## Internals
@@ -358,14 +344,6 @@ Base.macroexpand
 Base.@macroexpand
 Base.@macroexpand1
 Base.code_lowered
-Base.@code_lowered
 Base.code_typed
-Base.@code_typed
-Base.code_warntype
-Base.@code_warntype
-Base.code_llvm
-Base.@code_llvm
-Base.code_native
-Base.@code_native
 Base.precompile
 ```
