+++ stdlib/InteractiveUtils/docs/src/index.md
@@ -1,0 +1,25 @@
+# Interactive Utilities
+
+```@docs
+InteractiveUtils.apropos
+InteractiveUtils.varinfo
+InteractiveUtils.versioninfo
+InteractiveUtils.methodswith
+InteractiveUtils.subtypes
+InteractiveUtils.edit(::AbstractString, ::Integer)
+InteractiveUtils.edit(::Any)
+InteractiveUtils.@edit
+InteractiveUtils.less(::AbstractString)
+InteractiveUtils.less(::Any)
+InteractiveUtils.@less
+InteractiveUtils.@which
+InteractiveUtils.@functionloc
+InteractiveUtils.@code_lowered
+InteractiveUtils.@code_typed
+InteractiveUtils.code_warntype
+InteractiveUtils.@code_warntype
+InteractiveUtils.code_llvm
+InteractiveUtils.@code_llvm
+InteractiveUtils.code_native
+InteractiveUtils.@code_native
+```
