+++ stdlib/Mmap/test/runtests.jl
@@ -194,22 +194,22 @@ close(s)
 s = open(file, "r")
 @test isreadonly(s)
 b = @inferred Mmap.mmap(s, BitArray, (17,13))
-@test Base._check_bitarray_consistency(b)
+@test Test._check_bitarray_consistency(b)
 @test b == trues(17,13)
 @test_throws ArgumentError Mmap.mmap(s, BitArray, (7,3))
 close(s)
 s = open(file, "r+")
 b = Mmap.mmap(s, BitArray, (17,19))
-@test Base._check_bitarray_consistency(b)
+@test Test._check_bitarray_consistency(b)
 rand!(b)
 Mmap.sync!(b)
 b0 = copy(b)
-@test Base._check_bitarray_consistency(b0)
+@test Test._check_bitarray_consistency(b0)
 close(s)
 s = open(file, "r")
 @test isreadonly(s)
 b = Mmap.mmap(s, BitArray, (17,19))
-@test Base._check_bitarray_consistency(b)
+@test Test._check_bitarray_consistency(b)
 @test b == b0
 close(s)
 finalize(b); finalize(b0)
@@ -223,7 +223,7 @@ end
 @test filesize(file) == 9
 s = open(file, "r+")
 m = Mmap.mmap(s, BitArray, (72,))
-@test Base._check_bitarray_consistency(m)
+@test Test._check_bitarray_consistency(m)
 @test length(m) == 72
 close(s); finalize(m); m = nothing; GC.gc()
 rm(file)
