+++ test/llvmcall2.jl
@@ -1,5 +1,7 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
+using InteractiveUtils: code_llvm
+
 function declared_floor(x::Float64)
     return ccall("llvm.floor.f64", llvmcall, Float64, (Float64,), x)
 end
